dne-paqx/src/test/java/com/dell/cpsd/paqx/dne/service/task/handler/preprocess/PingIdracTaskHandlerTest.java
/**
 * Copyright &copy; 2017 Dell Inc. or its subsidiaries.  All Rights Reserved. Dell EMC Confidential/Proprietary Information
 *
 */

package com.dell.cpsd.paqx.dne.service.task.handler.preprocess;

import com.dell.cpsd.paqx.dne.domain.Job;
import com.dell.cpsd.paqx.dne.service.NodeService;
import com.dell.cpsd.paqx.dne.service.model.Status;
import com.dell.cpsd.paqx.dne.service.model.TaskResponse;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import static org.junit.Assert.*;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.doReturn;
import static org.mockito.Mockito.never;
import static org.mockito.Mockito.spy;
import static org.mockito.Mockito.verify;

/**
 * The tests for PingIdracTaskHandler.
 *
 * <p>
 * Copyright &copy; 2017 Dell Inc. or its subsidiaries. All Rights Reserved. Dell EMC Confidential/Proprietary Information
 * </p>
 *
 * @since 1.0
 */
@RunWith(MockitoJUnitRunner.class)
public class PingIdracTaskHandlerTest
{
    @Mock
    private NodeService service;

    @Mock
    private Job job;

    @Mock
    private TaskResponse response;

    private PingIdracTaskHandler handler;
    private PingIdracTaskHandler spy;

    private String taskName = "pingIdracTask";
    private String stepName = "pingIdracStep";

    /**
     * The test setup.
     *
     * @since 1.0
     */
    @Before
    public void setUp()
    {
        this.handler = new PingIdracTaskHandler(this.service);
        this.spy = spy(this.handler);
    }

    @Test
    public void executeTask() throws Exception
    {
        doReturn(this.response).when(this.spy).initializeResponse(this.job);

        assertTrue(this.spy.executeTask(this.job));
        verify(this.response).setWorkFlowTaskStatus(Status.SUCCEEDED);
        verify(this.response, never()).addError(anyString());
    }

}